move instantiated enums to appropriate components
