+++ TODO.md
@@ -1,0 +1,7 @@
+# TODOs
+
+## Better VSCode
+
+* Code Actions (e.g., for abstract classes)
+* isort
+*
