+++ components/niagads/enums/core.py
@@ -22,27 +22,4 @@ class CustomStrEnum:
         """ return a list of all values in the StrEnum """
         return [e.value for e in self]
     
-    
-class ClassProperties(CustomStrEnum, UppercaseStrEnum):
-    """enum for functions that extract info about a class -- methods or members?
-    """
-    METHODS = auto()
-    MEMBERS = auto()
-    
-class ErrorActions(CustomStrEnum, UppercaseStrEnum):
-    """enum for possible behaviors on critical errors
-    """
-    IGNORE = auto()
-    FAIL = auto()
-    WARN = auto()
-    
-class RegularExpressions(CustomStrEnum, StrEnum):
-    """
-    commonly used regexps
-    """
-    PUBMED_ID = r'^([0-9]{8}|PMID:[0-9]{8})$'
-    MD5SUM = r'^[a-fA-F0-9]{32}$'
-    DOI = r'(10[.][0-9]{4,}(?:[.][0-9]+)*\/(?:(?!["&\'<>])\S)+)$'
-    FILE_SIZE = r'^[.0-9]+\s?(K|M|G)$'
-    KEY_VALUE_PAIR = r'^[^\s=]+=[^=;]+$' # key=value
