+++ components/niagads/json_validator/format_checkers.py
@@ -12,7 +12,7 @@ from jsonschema import FormatChecker
 from re import RegexFlag
 
 from niagads.string.core import matches
-from niagads.enums.core import RegularExpressions as RE
+from niagads.string.regular_expressions import RegularExpressions as RE
 
 JSONSchemaFormatChecker = FormatChecker()
