+++ components/niagads/string/__init__.py
@@ -1,3 +1,4 @@
 from niagads.string import core
+from niagads.string.regular_expressions import RegularExpressions
 
-__all__ = ["core"]
+__all__ = ["core", "RegularExpressions"]
